refactor(constants, message): 상수 분리 및 네이밍 수정, 반영
