+++ src/constant/constants.js
@@ -1,0 +1,32 @@
+export const SYMBOL = {
+  one: 1,
+  three: 3,
+  four: 4,
+  five: 5,
+  six: 6,
+  hundred: 100,
+  delimiter: ',',
+  input_prefix: '>',
+  space: ' ',
+};
+
+export const REGEXP = {
+  numericPattern: /^\d+$/,
+};
+
+export const PRIZE = {
+  first: 2_000_000_000,
+  second: 30_000_000,
+  third: 1_500_000,
+  fourth: 50_000,
+  fifth: 5_000,
+};
+
+export const LOTTO_RULES = {
+  length: 6,
+  min_number: 1,
+  max_number: 45,
+  cost: 1_000,
+  restart: 'y',
+  stop: 'n',
+};
