+++ src/utils/random.js
@@ -1,13 +1,10 @@
-import { CONSTANTS } from '../constant/index.js';
-
 const Random = {
   pickUniqueNumbersInRange({ minNumber, maxNumber, count }) {
-    if (count > maxNumber - minNumber + CONSTANTS.one)
-      throw new Error('Cannot generate unique numbers. Count exceeds range.');
+    if (count > maxNumber - minNumber + 1) throw new Error('Cannot generate unique numbers. Count exceeds range.');
     const uniqueNumbers = new Set();
 
     while (uniqueNumbers.size < count)
-      uniqueNumbers.add(Math.floor(Math.random() * (maxNumber - minNumber + CONSTANTS.one)) + minNumber);
+      uniqueNumbers.add(Math.floor(Math.random() * (maxNumber - minNumber + 1)) + minNumber);
 
     return Array.from(uniqueNumbers);
   },
