+++ src/view/InputView.js
@@ -1,8 +1,9 @@
+import { SYMBOL } from '../constant/constants.js';
+import { INPUT_MESSAGES } from '../constant/messages.js';
 import Lotto from '../domain/lotto.js';
 import WinningLotto from '../domain/winningLotto.js';
 import ReadLine from '../utils/readLineAsync.js';
 import { validateCost, validateRestartResponse } from '../utils/validation.js';
-import { INPUT_MESSAGES, CONSTANTS } from '../constant/index.js';
 
 const InputView = {
   async readCost() {
@@ -19,7 +20,7 @@ const InputView = {
   async readLotto() {
     try {
       const numbers = (await ReadLine.readLineAsync(INPUT_MESSAGES.winningNumber))
-        .split(CONSTANTS.comma)
+        .split(SYMBOL.delimiter)
         .map((number) => Number(number));
 
       return new Lotto(numbers);
